backend/nw_backend/requirements.txt

django-environ==0.11.2


Django==4.2.13
wagtail==6.1.1
wagtail-seo==2.5.0
dj-database-url==2.1.0
websockets==12.0
django-cors-headers==4.3.1
django-filter==24.2
django-firebase-auth==1.2.1
django-modelcluster==6.3
django-permissionedforms==0.1
django-storages==1.14.3
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.1
psycopg2-binary==2.9.9
l18n==2021.3
fastapi==0.111.0
gunicorn==22.0.0
whitenoise==6.6.0
uvicorn==0.30.1